solving working time issues and only the halaqah of a markej show up

+++ src/Hafiz.Application/DTO/Attendance/SaveAttendanceDto.cs
@@ -12,6 +12,6 @@ namespace Hafiz.DTOs
         public Guid ClassID { get; set; }
         public AttendanceStatus Status { get; set; }
         public DateTime Date { get; set; }
-        public int Hours { get; set; }
+        public float Hours { get; set; }
     }
 }
